Revert "Rename build/make/target/board/Android.mk to android-inf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e03b77e-43d5-4794-aafd-b381cc093e99). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377781876
Change-Id: I86819f4f35c7f6a0000021b1793aeeab4932e5e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com