Revert "Rename build/make/target/board/Android.mk to android-inf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0dfe7722-2331-4547-b092-c4b5a8fca92f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377781799
Change-Id: I29d5b222c8a5abb7dddb40b87fffc989ce7b89d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com