Revert "Rename build/make/target/board/Android.mk to android-inf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83d0264b-efff-4999-a7d4-313ab0312d1a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377781521
Change-Id: Ia762b6908adb229478e8d3a3515ba6015f8756f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com