Revert "Rename build/make/target/board/Android.mk to android-inf..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36919618-0ce4-428d-91ba-0cb5879c3670). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377840339
Change-Id: Ifd20cb7e744b722bce48712a05ab2e42db4a80b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com