Revert "Export vbmeta variables to soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5ba07ab2-430e-44f2-a4e3-0d15d6000e77). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377985499
Change-Id: Ie0ae68d075ff81d7acc146a5e9561c711e60f603
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com