Revert "Export vbmeta variables to soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f970c94b-5bc2-4a7d-bb64-f2156b4a07b1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378000034
Change-Id: I6afc05a822efc52c3c31cf0fdecb1f39d4905954
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com