Revert "Revert^2 "Rename build/make/target/board/Android.mk to a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4477047b-8eda-4e85-9b23-5c2e97523b5a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378024079
Change-Id: I9016e44344b1491883392d35da9cf4a7858e921a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com