Revert "Revert^2 "Rename build/make/target/board/Android.mk to a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e20f24b7-84c5-42b7-8bba-8657250fb0e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378023950
Change-Id: Id43a552e1bdf4ddce5b28524a4798e522bbfa0b1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com