Revert "Revert^2 "Rename build/make/target/board/Android.mk to a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c081da69-bf74-461e-b4a5-64361c7c3a7d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378024684
Change-Id: Icffd8c2937357b7296aa11fd77798b715cd6fe5d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com