Revert "Revert^2 "Rename build/make/target/board/Android.mk to a..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b151589-ebee-431c-9723-905b4f5dd9cc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 378021131
Bug: 378024873
Change-Id: Ie188d182ad37749d411c666aa889f58e79dabd9d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com