Revert "Check remaining bytes before attempting to read from buffer"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b7eeae0-7d1e-4c50-920a-3459690a30d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22606caef04bddf41c73fc43bb78128e38134be4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com