Revert "Implement test discovery agent for test zip discovery"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b7eeae0-7d1e-4c50-920a-3459690a30d7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0690a32b4d21f389a15eb68aa85a950a63857d2f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com