Revert "Write list of files in the ramdisk"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5dacdfe0-f409-4b5c-af6e-80077a62dc2f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9094b340956422b592f296fe19cf140311178ba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com