Revert "Export some board config vars to autogenerate android_in..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6426791-18c6-4f27-8e51-58c6eed1cc57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaadcb42a90c70da6dfc426017a845fa270abe750
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com