Make CTS build work without BREAKPAD_GENERATE_SYMBOLS=false
Summary: Without this change building with Breakpad symbols causes a build breakage.
The breakage occurs because the Breakpad dump_syms program gives an error code if given a
library (*.so file) which does NOT contain debug symbols. This change adapts an existing test
that the input is an ELF file to also verify that it contains a section called .debug_info.

Test: m cts
Check that sym files are still generated:
find out -name '*.so.sym' -ls | more

Change-Id: I6f213555256238dd470c0be5f570fc6cb380dfc0
Signed-off-by: Abdelrahman Daim <adaim@meta.com>