Revert "Build ramdisk's build.prop with soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a0562dc-37f7-4757-97c4-6fca47515da7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f42aef64453e160311830544417922d620136ee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com