Revert "Set system property from list of backported fixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80881fb5-6d37-4a4a-bfdf-31f75ef99bb4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379198534
Change-Id: I3d0bac0c4f4389c72c4a6d88d5add2c63567a787
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com