Add multiprocessing to sign_target_files_apks.py.
Summary: Use the multiprocessing module to run both APK and APEX signing steps as subprocesses to speed up signing.
Note that while the signing is done in a separate process, the writing of the signed APK/APEX is done
back on the main process.
Implemented by creating common.TaskQueue,  which is a queue of tasks where work can to be split into
two steps, where the first  step can be done in a separate process, and then the results of that work are
processed in a second step in the main process.
Checking the queue inside the for loop is done to reduce the total memory stored, since APKs are returned
in memory from the signing process. This is not a concern with signed APEXs since the returned value is just a name
of a file with the signed APEX.
Signing in subprocesses reduces signing time from  16 min 5 sec to  11 min 17 sec for a 30% reduction in signing speed.

Test: Run sign_target_files_apks with and without --multiprocessing

Change-Id: I58c52aae63992b310057a5f343364fc4c1f10a39
Signed-off-by: Abdelrahman Daim <adaim@meta.com>