Apply common.AsyncTask to add_img_to_target_files.py
Summary: Apply common.AsyncTask to add_img_to_target_files.py, which will also speed up signing.
Break the implementation of CreateImage into two closures that are then passed
to the TaskQueue.  The first closure builds the image in a separate process, and the
second closure adds the image to the zip file in the main process.
Along with the first Change Request in this stack, signing time is reduced 50% from 16m to 8m.

Test: Run sign_target_files_apks with and without --multiprocessing

Change-Id: If5cee5b978145e04989cd09ffe2c29a44318654d
Signed-off-by: Abdelrahman Daim <adaim@meta.com>