Revert "Remove special case for ltp build rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b0f3110-0a7f-429e-9722-37e32f4afdba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379763500
Change-Id: I3a2f6c0ca7f0f4cfbd8e7329ce3948566987a37f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com