Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6407b730-526d-4bc9-beb0-5296010a07c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379805526
Change-Id: I101966babdc7ca3b52b92ddd09630a0684ff6c84
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com