Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/303d4734-31c3-4e37-a218-731345c2cc26). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379804162
Change-Id: Icbe7fdb7bd2f3ed66b37290fe40132541c7f276e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com