Revert "Export ProductBuildVendorBootImage to soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6f06601-41ae-4d7a-b476-1fcc7ce9b7a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379806999
Change-Id: I94f13b77f640373b4548bc139f7f703e1365f92a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com