Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6b0d7de-f636-4262-8749-62afc7950f92). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379869783
Change-Id: I058938207f1ddcf0549386a0d13b97812466677d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com