Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/08192585-078d-44ea-b32e-921a0e7c51ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379934185
Change-Id: I8483980ff641c2cd08f0a555073f2d02a035393d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com