Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d6ef5e4b-ea18-4c9f-bb5e-ffa3f5a56ee0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379985927
Change-Id: I9a98a3b75c71249785114cc0bc3a826fa1a8d358
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com