Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6974d67a-9ad1-4b34-a7e9-0197c551b284). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380125921
Change-Id: I2bb2f6359fa72d6b3d879699a87391828b419627
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com