Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/075fb89c-42db-4f3b-9115-21e017647140). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380125844
Change-Id: I3f5e0f70611c6851c16ffa0eab5eae22d8779226
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com