Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/73169f3a-e150-4d32-bf2f-6f50b447a367). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380106417
Change-Id: I2fc20003ad527d7c37b6646476d9b8c6af1d290d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com