Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c0164ae-fda6-4b0b-b781-2079fb84a552). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380131315
Change-Id: I3b5d42ebb27f9bc9177aa3b4c3273eb7c53a5b41
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com