Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c64fc96f-cc03-4a22-af40-3bcb07f5a9bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380108094
Change-Id: Ic26ced15634dc9b6c99945de94d627fd5c6c9f8e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com