Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a005ab4a-ac50-471a-80b6-99fca8df6950). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380332248
Change-Id: Ibe7e85a1d80373aae9ede67039a90e67fd231e3c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com