Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b5c78a7-60b4-49ad-b524-f24bc0fa6991). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380340682
Change-Id: Ic37a0fd4890a2787fdc304a6b24c08b7d39cd14d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com