Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20f0060a-226b-4aec-96a3-5c61a1c01fc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380331962
Change-Id: I7549834b252e54768fa89ca26a95ef343c960b97
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com