Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0143200-8d7f-41ce-8e71-158155c08450). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380340289
Change-Id: I138e70a6ea3a2b93c69512f7a76d6aed3a20b6c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com