Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f946863-b13d-40ec-82a4-7344c435c7d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380340231
Change-Id: I4977cecb877f263adb303ecafaa6c0349cc448bb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com