Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d028928-07c7-4be3-b7b9-c81437419c46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380331785
Change-Id: Ie6fafe31453262802f2bfc0ca93aeed0e9c290c8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com