Revert "Export BOARD_INCLUDE_DTB_IN_BOOTIMG to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f376422c-648b-4ee2-9d99-dfacbde3997b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 380339751
Change-Id: Iad4f9fc89e979e9eb718cc4a0b6ecba59634f4e8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com