Salt init_boot and vendor_boot using the build number
If you don't provide a salt, avbtool uses random bytes as the salt.
The randomness is bad for build caching and being able to diff
partitions to ensure code changes aren't breaking anything.

It's not exactly clear if the salt is necessary at all or not, but
as an intermediate step, salt the images using the build number instead
of purely random bytes.

I've only done the init_boot and vendor_boot partitions in this cl,
but I plan to expand it to all the other partitions. There are other
diffs besides the avb hash in those partitions right now though so it's
harder to test.

Bug: 380508890
Test: m soong_generated_init_boot_filesystem_test
Change-Id: I56289509497a2f0d500ffbbd677517fa0f748340