Revert "update java local flag read code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57473ec5-f00e-43aa-ad13-1592bb8ff4a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia16389b6d38d59bbcb2b83d1bc0c447cf97bcaae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com