Revert "update java local flag read code"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4525cb3-586b-47fb-bfc3-73cedbf38913). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iea618cc8a6111fcbbc2a7d3139ac284e906075a8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com