Revert "Export boot image-specific variables to soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3b86eaf-f648-4a96-bd37-e1dabced983b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85820d5e6f54a8c00391ed9566d72cdce1f68c24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com