Revert "Build ramdisk's build.prop with soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2d087692-59d4-45a9-93c7-cc3d8dd55a23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic282c0904e8b1d691555f81231fec793d0552036
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com