Very early test for test_discovery logic for general-tests.zip
The get_build_targets_impl should be mostly ok, I haven't been able to test it. I'm struggling a lot with get_package_outputs_commands_impl. It's essentially a copy of what happens in general-tests.mk, but with the extra step of filtering down COMPATIBILITY.general-tests.FILES down based on the modules returned form the test_discovery agent. I don't have the time to fully replicate the logic in the script. I'm wondering if it would be possible to just pull out that part of the makefile into just a bash script (since that's what it is). All the environment variables are available via the soong dumpvars that i do early on in that script, and you'd just replace the initial `echo $(sort $(COMPATIBILITY.general-tests.FILES)` line with something filtered down from the `general-tests-files-list` output file. Will discuss. Note: changes to the general-tests.mk rule should be included in the previous CL, not this one.

Change-Id: Ia304e9702750c468c83ef82337c73a3862c044b9