Set `TARGET_RECOVERY_UI_LIB` as a Soong config string list
`TARGET_RECOVERY_UI_LIB` should be set as a string list using
`soong_config_set_string_list`, not as a string, because it usually
specifies multiple libraries.

Note: Currently, due to the issue that the `select` syntax never
reaches the `default[]` case when processing `string_list`, the default
value of `TARGET_RECOVERY_UI_LIB` needs to be set here first.

Bug: 375341427
Bug: 381317459
Test: m -j librecovery_ui_ext
Change-Id: I90f6d6e4d591f401ccac7a23100a5b2e1a5db639