Revert "remove file check"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e0333ab-af6e-45ac-9bc7-09ad6bf32f3d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 379975903
Change-Id: I96ab32cc1365496de7061182ad76a4b8daacc1d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com