Revert "Handle null exception messages in FeatureFlagsImpl.java...."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/571f99d2-3b7d-4ea5-8e1e-23bbfc2e6883). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89d0ed94266a1062090ec29d66ea2fbdbd838b24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com