Revert "Export recovery partition related variables to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/56808509-0eab-483e-a7d7-df526ed8db2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381962754
Change-Id: I67a77bb644049152274e2381dda8fced57a749f5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com