Revert "Export recovery partition related variables to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9dba5629-e820-49cc-973f-82d8fb6eb295). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381965574
Change-Id: Id6e6987472965b61613a86c7588a86f24eba2b1a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com