Revert "Export recovery partition related variables to Soong"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b5e35c2-e6c4-46eb-a8d0-38d8b96ea05e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 381967742
Change-Id: Ia18716847a3c607c69f5f04315076a7801dceaad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com