Fix a race condition in edit monitor
When there are multiple edit monitor starting around the same time,
there will be cases that the pidfile is deleted or overridden as
multiple processes are trying to remove and create the pidfile around
the saem time (see details in b/382135550).
This cl fix it by checking the pid in the pidfile and
only remove the pidfile if the pid stored there is the expected one.

Bug: 382135550
Test: Manual
Change-Id: Ifdb4c29c8b36052fea83ce59de1be9d99db3a852