Add fingerprint check to rust codegen.
Will only run if enable_fingerprint is true, which we will set when
version >= 2. In other words, the flag guarding here is the same build
flag guard that we have nothing to set the version to 2, so this code
won't run.

If the fingerprint doesn't match, return default! Only do this for new
storage (allow_instrumentation=true) as for old storage there are no
files, so no fingerprint.

Also, I pulled the expected code into separate files rather than inline
string constants because I was having trouble reading them.

Test: added test case for when fingerprint is enabled
Change-Id: I3b930e8a1368037f9ea6a32c1df3d567f77a04cd