Revert "Filter read-only + disabled flags from aconfig_flags.pb"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c89ff8de-bac5-413a-8a1b-03f4d5cf1f23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382295646
Change-Id: Iba75b314c01780c17cbd3391467f5ff52c8b44a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com