Revert "Filter read-only + disabled flags from aconfig_flags.pb"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fb9699c0-c7ef-4a08-821a-e8f5d6f6c810). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382294819
Change-Id: I62e41ca756b4ba6ae755bae60a7feb3df59f7872
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com