Revert "Filter read-only + disabled flags from aconfig_flags.pb"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3958ac7c-4d10-4a04-8ebe-a963a814212e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbada834e196ada875d07444cbb93eb4cf5b8bce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com