Revert "Add GSI targets using soong built system image"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/184f3e2b-0e0c-49f7-90cc-b7c4c2a4c49b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5013769cc56d9eec037c326a18bd36dda47eb9e4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com