Revert "Return default value if fingerprint doesn't match"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f944315-e4dc-494c-9999-0eecdaeaca8c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 382606466
Change-Id: Ic161aa624dd047d0b9e2e2a2b8a604f58f6efcf0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com